Moved reusable code, minor cleanup. Gave credit for original project in readme file.
