+++ README.md
@@ -1,1 +1,5 @@
 # ModemDotNet
+
+This project implements Y Modem in C# as a .NET Core 2.0 class library.
+
+Work in this project is heavily based on an adaptation from https://github.com/aesirot/ymodem.
